app: change GimpSettings to store its time as gint64
and use GDateTime for handling it.

+++ app/core/gimpsettings.h
@@ -39,7 +39,7 @@ struct _GimpSettings
 {
   GimpViewable  parent_instance;
 
-  guint         time;
+  gint64        time;
 };
 
 struct _GimpSettingsClass
